google-functions/preprocess-functions/requirements.txt
requests==2.25.1
google-cloud-firestore>=2.1.0
python-dotenv>=0.19.0